Reposition repository for pgeu-website
This commit removes all "shared parts" of the pgeu-system, leaving those
in the repository for the actual system.

As of this commit, this repository only has the functionality that is
unique to PostgreSQL Europe, and not used by other users of the system.
This includes all the actual webpage contents, logos etc, and in
particular it includes the application "cmutuel", as this bank
integration is highly unlikely to ever be used by somebody else (and
serves as a good example for how to put a code module in the separate
repository)

This also moves the pgeu specific media files to /media/local/, to make
it possible to serve up this separately from the shared media directory
in an efficient way from the webserver.