Move cm data parsing into payment provider
This moves the guts of the parsing data over into the payment provider,
leaving the nightly batchjob to just call it. This is in preparation of
changes in pgeu-system that will support manual uploading of files with
bank transactions.

In light of the same preparation, also add instructions for how to
manually download the transaction list from cm. Not used today, but will
be by the new code once ready.