Migrate cmutuel parsing code to new upstream framework
Instead of writing code for each bank output file format this framework
allows it to just be a JSON specification file. Which can then live in
the upstream repo itself since it's even less likely to change.

In doing so, create a migration that changes the existing
InvoicePaymentMethod to use the new framework (while copying the config
of it), and migrate all the existing downloaded transactions into the
new format.

Intentionally leave the old table around in the database for now, in
case something goes wrong and we need forensics :) At a later date we'll
add a migration to drop it.